Move most of the option setting code out of main()
