+++ src/config.rs
@@ -19,8 +19,13 @@
  *                                                                      *
  ************************************************************************/
 
+use ruleset::KgsChinese;
 use ruleset::Minimal;
 use ruleset::Ruleset;
+
+use getopts::Matches;
+use std::ascii::OwnedAsciiExt;
+use std::process::exit;
 
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub struct UctConfig {
@@ -97,6 +102,89 @@ impl Config {
                 tuned: true,
             },
         }
+    }
+
+    pub fn set_from_opts(&mut self, matches: &Matches) {
+        if matches.opt_present("empty-area-prior") {
+            let arg = matches.opt_str("empty-area-prior").unwrap();
+            self.uct.priors.empty = match arg.parse() {
+                Ok(v) => v,
+                Err(_) => {
+                    println!("Unknown value ({}) as argument to --empty-area-prior", arg);
+                    exit(1);
+                }
+            }
+        }
+
+        if matches.opt_present("use-atari-check-in-playouts") {
+            let arg = matches.opt_str("use-atari-check-in-playouts").map(|s| s.into_ascii_lowercase()).unwrap();
+            config.playout.atari_check = match arg.parse() {
+                Ok(v) => v,
+                Err(_) => {
+                    println!("Unknown value ({}) as argument to --use-atari-check-in-playouts", arg);
+                    exit(1);
+                }
+            }
+        }
+
+        if matches.opt_present("use-empty-area-prior") {
+            let arg = matches.opt_str("use-empty-area-prior").map(|s| s.into_ascii_lowercase()).unwrap();
+            self.uct.priors.use_empty = match arg.parse() {
+                Ok(v) => v,
+                Err(_) => {
+                    println!("Unknown value ({}) as argument to --use-empty-area-prior", arg);
+                    exit(1);
+                }
+            }
+        }
+
+        if matches.opt_present("use-ladder-check-in-playouts") {
+            let arg = matches.opt_str("use-ladder-check-in-playouts").map(|s| s.into_ascii_lowercase()).unwrap();
+            self.playout.ladder_check = match arg.parse() {
+                Ok(v) => v,
+                Err(_) => {
+                    println!("Unknown value ({}) as argument to --use-ladder-check-in-playouts", arg);
+                    exit(1);
+                }
+            }
+        }
+
+        let reuse_subtree_arg = matches.opt_str("reuse-subtree").map(|s| s.into_ascii_lowercase());
+        let reuse_subtree = match reuse_subtree_arg {
+            Some(arg) => {
+                match arg.parse() {
+                    Ok(v) => v,
+                    Err(_) => panic!("Unknown value ({}) as argument to --reuse-subtree", arg)
+                }
+            },
+            None => true
+        };
+        let log = matches.opt_present("l");
+
+        let threads = match matches.opt_str("t") {
+            Some(s) => {
+                match s.parse() {
+                    Ok(n)  => n,
+                    Err(_) => 1
+                }
+            },
+            None => 1
+        };
+        let rules_arg = matches.opt_str("r").map(|s| s.into_ascii_lowercase());
+        let ruleset = match rules_arg {
+            Some(r) => Ruleset::from_string(r),
+            None    => KgsChinese
+        };
+
+        let policy = matches.opt_str("P").map(|s| s.into_ascii_lowercase());
+        self.log = log;
+        self.ruleset = ruleset;
+        self.threads = threads;
+        self.uct.tuned = match policy {
+            Some(str) => if str == "ucb1" { false } else { true},
+            _ => true
+        };
+        self.uct.reuse_subtree = reuse_subtree;
     }
 
 }
